Create README.md
This project is focused on developing a machine learning model to predict the chance of diabetes in hospital patients. Out of 9 models developed, CatBoost classification model had the highest accuracy and precision of 96.4%. For prediction with proabability from 65% and above, no confirmatory test was recommended. However, for prediction with probability below 60%, a confirmatory test was recommended.